requirements.txt
Django
Pillow
django-bootstrap4
django-widgets-improved
djangorestframework
django-filter